requirements.txt
google-genai
vnstock
pandas
numpy
fastapi
uvicorn
requests
python-dotenv
openpyxl
python-dotenv
python-dateutil